+++ authentication/authenticate.go
@@ -10,7 +10,9 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/joinself/self-go-sdk/pkg/kidhelper"
 	"github.com/joinself/self-go-sdk/pkg/ntp"
+	"github.com/joinself/self-go-sdk/pkg/request"
 	"github.com/joinself/self-go-sdk/pkg/siggraph"
 	"github.com/lucasb-eyer/go-colorful"
 	"github.com/skip2/go-qrcode"
@@ -292,7 +294,7 @@ func (s *Service) authenticationResponse(selfID string, resp []byte) (string, er
 		return cid, err
 	}
 
-	kid, err := getJWSKID(resp)
+	kid, err := kidhelper.GetJWSKID(resp)
 	if err != nil {
 		return cid, err
 	}
@@ -322,7 +324,7 @@ func (s *Service) authenticationPayload(cid, selfID string, exp time.Duration) (
 		selfID = "-"
 	}
 
-	req := map[string]string{
+	req := map[string]interface{}{
 		"jti":       uuid.New().String(),
 		"cid":       cid,
 		"typ":       RequestAuthentication,
@@ -334,28 +336,7 @@ func (s *Service) authenticationPayload(cid, selfID string, exp time.Duration) (
 		"device_id": s.deviceID,
 	}
 
-	payload, err := json.Marshal(req)
-	if err != nil {
-		return nil, err
-	}
-
-	opts := &jose.SignerOptions{
-		ExtraHeaders: map[jose.HeaderKey]interface{}{
-			"kid": s.keyID,
-		},
-	}
-
-	signer, err := jose.NewSigner(jose.SigningKey{Algorithm: jose.EdDSA, Key: s.sk}, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	signature, err := signer.Sign(payload)
-	if err != nil {
-		return nil, err
-	}
-
-	return []byte(signature.FullSerialize()), nil
+	return request.Serialize(req, s.keyID, s.sk)
 }
 
 // builds a list of all devices associated with an identity
@@ -378,38 +359,4 @@ func (s Service) paidActions() bool {
 	}
 
 	return app.PaidActions
-}
-
-func getKID(token string) (string, error) {
-	data, err := base64.RawURLEncoding.DecodeString(strings.Split(token, ".")[0])
-	if err != nil {
-		return "", err
-	}
-
-	hdr := make(map[string]string)
-
-	err = json.Unmarshal(data, &hdr)
-	if err != nil {
-		return "", err
-	}
-
-	kid := hdr["kid"]
-	if kid == "" {
-		return "", errors.New("token must specify an identifier for the signing key")
-	}
-
-	return kid, nil
-}
-
-func getJWSKID(payload []byte) (string, error) {
-	var jws struct {
-		Protected string `json:"protected"`
-	}
-
-	err := json.Unmarshal(payload, &jws)
-	if err != nil {
-		return "", err
-	}
-
-	return getKID(jws.Protected)
 }
